refactor: PR review changes
http4s modules renamed to http4s-blaze-* to be more precise

refactor: CorrelationIdMiddleware made more general

refactor: Http4sRouting made more general
