--- http4s-server/src/main/scala/com/avast/server/toolkit/http4s/Http4sServerModule.scala
@@ -1,39 +1,0 @@
-package com.avast.server.toolkit.http4s
-
-import java.net.{InetSocketAddress, StandardSocketOptions}
-
-import cats.effect.{ConcurrentEffect, Resource, Timer}
-import org.http4s.HttpApp
-import org.http4s.server.Server
-import org.http4s.server.blaze.BlazeServerBuilder
-
-import scala.concurrent.ExecutionContext
-import scala.concurrent.duration.Duration
-import scala.language.higherKinds
-
-object Http4sServerModule {
-
-  /** Makes [[org.http4s.server.Server]] initialized with the given config and [[org.http4s.HttpApp]]. */
-  def make[F[_]: ConcurrentEffect: Timer](config: Http4sServerConfig,
-                                          httpApp: HttpApp[F],
-                                          executionContext: ExecutionContext): Resource[F, Server[F]] = {
-    BlazeServerBuilder[F]
-      .bindSocketAddress(InetSocketAddress.createUnresolved(config.listenAddress, config.listenPort))
-      .withHttpApp(httpApp)
-      .withExecutionContext(executionContext)
-      .withoutBanner
-      .withNio2(config.nio2Enabled)
-      .withWebSockets(config.webSocketsEnabled)
-      .enableHttp2(config.http2Enabled)
-      .withResponseHeaderTimeout(Duration.fromNanos(config.responseHeaderTimeout.toNanos))
-      .withIdleTimeout(Duration.fromNanos(config.idleTimeout.toNanos))
-      .withBufferSize(config.bufferSize)
-      .withMaxRequestLineLength(config.maxRequestLineLength)
-      .withMaxHeadersLength(config.maxHeadersLength)
-      .withChunkBufferMaxSize(config.chunkBufferMaxSize)
-      .withConnectorPoolSize(config.connectorPoolSize)
-      .withChannelOption[java.lang.Boolean](StandardSocketOptions.TCP_NODELAY, config.socketOptions.tcpNoDelay)
-      .resource
-  }
-
-}
